python_100_days/pyproject.toml
[tool.poetry]
name = "python-100-days"
version = "0.1.0"
description = ""
authors = ["Robin Burri <robin@Robins-MacBook-Air.local>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
requests = "^2.31.0"
twilio = "^8.5.0"
python-dotenv = "^1.0.0"
bs4 = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
